+++ streamndr/model/__init__.py
@@ -1,8 +1,10 @@
 """Novelty detection models"""
 from .minas import Minas
+from .ecsminer import ECSMiner
 from .ecsminerwf import ECSMinerWF
 
 __all__ = [
     "Minas",
+    "ECSMiner",
     "ECSMinerWF",
 ]
