+++ streamndr/utils/__init__.py
@@ -8,4 +8,5 @@ __all__ = [
     "MCIKMeans",
     "MicroCluster",
     "ShortMemInstance",
+    "ClusterModel",
 ]
